requirements.txt
Django>=5.2.4
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
requests>=2.32.4


